refactor: move helper functions under common, in usage specific named files
