+++ pkg/api/controller.go
@@ -178,7 +178,7 @@ func (c *Controller) Run(reloadCtx context.Context) error {
 		if c.Config.HTTP.TLS.CACert != "" {
 			clientAuth := tls.VerifyClientCertIfGiven
 			if (c.Config.HTTP.Auth == nil || c.Config.HTTP.Auth.HTPasswd.Path == "") &&
-				!anonymousPolicyExists(c.Config.HTTP.AccessControl) {
+				!c.Config.HTTP.AccessControl.AnonymousPolicyExists() {
 				clientAuth = tls.RequireAndVerifyClientCert
 			}
